+++ src/common/utils/orgConfigUtils.ts
@@ -4,7 +4,8 @@ import { glob } from 'glob';
 import puppeteer from 'puppeteer-core';
 import * as yaml from 'js-yaml';
 import { uxLog } from './index.js';
-import { SfError } from '@salesforce/core';
+import { Connection, SfError } from '@salesforce/core';
+import { DescribeSObjectResult } from '@jsforce/jsforce-node';
 
 const listViewRegex = /objects\/(.*)\/listViews\/(.*)\.listView-meta\.xml/gi;
 
@@ -152,4 +153,20 @@ export async function listMajorOrgs() {
     majorOrgs.push(props);
   }
   return majorOrgs;
+}
+
+export async function checkSfdxHardisTraceAvailable(conn: Connection) {
+  let traceObject: DescribeSObjectResult;
+  try {
+    traceObject = await conn.sobject("SfdxHardisTrace__c").describe();
+  } catch (e) {
+    throw new SfError("You need a List Custom Setting named SfdxHardisTrace__c, with Type__c and Key__c fields (both string,80)");
+  }
+  const traceObjectFields = traceObject.fields;
+  if (traceObjectFields.filter(field => field.name === "Type__c").length === 0) {
+    throw new SfError("You need a field Type__c (string,80) on SfdxHardisTrace__c in target org");
+  }
+  if (traceObjectFields.filter(field => field.name === "Key__c").length === 0) {
+    throw new SfError("You need a field Key__c (string,80) on SfdxHardisTrace__c in target org");
+  }
 }